I made a prototype of the melee termite that I think feels pretty good to fight.
It uses several timers and pauses animation since we don't really have a tired or stop animation.
It uses kinematic based motion to do the dash.
I changed the ground effect to be able to use an alternative knockback because it made more sense for the termites to knock the player back.
I changed the sprite flip code as I think it needed to work before the attack in a way that didn't just rely on move direction.