Prototype spider enemy
The spider now has a dash animation for when it dodges.
It now uses a spin animation and a large warning circle when it attacks.
It now has a lunge animation for when it jumps at the player.
Changed spider code a bit to account for discrepancies between it and other enemies.